beta defineProps
