+++ src/index.ts
@@ -10,6 +10,8 @@ import { oif } from "./package/reactive/oif";
 import { Orve, context } from "./package/default";
 import { Node, Fragment } from "./package/jsx";
 import { createAppRoot } from "./package/dom/root";
+import { defineProps } from "./package/dom/defineProps";
+
 export {
   createApp,
   refL,
@@ -24,7 +26,8 @@ export {
   Fragment,
   oif,
   context,
-  createAppRoot
+  createAppRoot,
+  defineProps
 };
 
 export default {
@@ -32,5 +35,6 @@ export default {
   ...Orve,
   Node,
   Fragment,
-  createAppRoot
+  createAppRoot,
+  defineProps
 };
